Replace buttons markup with actual <button> instead of <a>
Since an <a> needs an href attribute, we needed to use a placeholder on it, by default: "#".
This was causing the side effect of the page scrolling back to the top whenever you clicked on a vote button.
Also, a <button> is definitely the better semantic option in this case.